Changed command name + Changed Try-catch in mvc-contracts:update command
