+++ tests/feature/commands/MvcUpdateContractTest.php
@@ -3,7 +3,6 @@
 namespace Tychovbh\Tests\Mvc\feature\commands;
 
 use Tychovbh\Mvc\Contract;
-use Tychovbh\Mvc\Http\Resources\ContractResource;
 use Tychovbh\Tests\Mvc\TestCase;
 
 class MvcUpdateContractTest extends TestCase
@@ -13,14 +12,13 @@ class MvcUpdateContractTest extends TestCase
      */
     public function itCanUpdateContracts()
     {
-        factory(Contract::class, 2)->create(['template' => 'contract', 'signed_at' => null]);
         $contract = factory(Contract::class)->make(['signed_at' => null, 'status' => Contract::STATUS_CONCEPT]);
         $contract->template = 'contract';
         $contract->save();
 
-        $this->artisan('mvc-contract:update')->expectsOutput('All contracts are updated');
+        $this->artisan('mvc-contracts:update')->expectsOutput('1 contracts updated');
 
-        $contract = Contract::find(3);
+        $contract = Contract::find($contract->id);
         $this->assertTrue($contract->status === 'signed' && $contract->signed_at);
     }
 }
